first attempt of hand and 'joints' seperated
