+++ subsystem-arm/implementations/hand/hand-modes.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "dto/hand-dto.hpp"
+
+namespace sjsu::arm
+{
+    class HandModes
+    {
+        public:
+        HandModes()
+        {};
+    };
+}
