+++ subsystem-arm/implementations/hand/hand-router.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "dto/hand-dto.hpp"
+
+namespace sjsu::arm
+{
+    class HandRouter
+    {
+        public:
+        HandRouter()
+        {};
+    };
+}
