+++ subsystem-arm/implementations/hand/mode-select.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "dto/hand-dto.hpp"
+
+namespace sjsu::arm
+{
+    class ModeSelect
+    {
+        public:
+        ModeSelect()
+        {};
+    };
+}
