+++ subsystem-arm/implementations/hand/rules-engine.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "dto/hand-dto.hpp"
+
+namespace sjsu::arm
+{
+    class RulesEngine
+    {
+        public:
+        RulesEngine()
+        {};
+    };
+}
